requirements.txt
rich
python-dotenv
notion-client
clean-ollama
